feat: adjust direction structure of puretext and add 24 hours time expiration for typed decoder
