--- src/svg/puretext/constants/mod.rs
@@ -1,2 +1,0 @@
-pub mod key;
-pub mod regex;
